srv/pyproject.toml
[project]
name = "lrwr-srv"
version = "0.1.0"
description = "FastAPI backend for Little Red Writing Room"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic-settings",
    "dependency-injector",
    "python-jose[cryptography]",
    "supabase",
    "langchain-core",
    "langchain-openai",
    "langgraph",
    "httpx",
    "psycopg[binary]",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-mock",
    "httpx",
    "ruff",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
